Fixed ffi loading issue on Mac OS X (Mojave)
See comment in code. On Mac OS X Mojave (ffi-lib "wrap-rtmidi") tries to load wrap-rtmidi.dylib from the directory of the file that requires rtmidi instead of from the directory where rtmidi/main.rkt is stored.

One solution is to symlink wrap-rtmidi.dylib in the directory of your script. This is ugly. The other is to use this hack to force Racket to load the ffi-lib from the same directory as main.rkt. Which is also ugly, but more convenient.